v3.5: Migrate core pricing and inventory money logic to BigDecimal
- Refactored Product price storage from double to BigDecimal
- Updated Product getters, setters, compareTo(), and employee discount logic to use BigDecimal operations
- Updated InventoryManager price-range filtering to accept BigDecimal min / max values and compare prices without floating-point math
- Updated FurnitureItem and FoodItem constructors so subtype pricing now matches the BigDecimal Product model
- Updated Main test/product creation flow to build item prices with BigDecimal values
- Completed the v3.5 money-type migration needed before deeper economy and checkout work in v4

+++ src/models/FoodItem.java
@@ -1,5 +1,7 @@
 package models;
 
+import java.math.BigDecimal;
+
 public class FoodItem extends Product {
 
     // Initialize Food specific variables
@@ -13,7 +15,7 @@ public class FoodItem extends Product {
     boolean hasFlavor;
 
     // Initializing Product
-    public FoodItem(String productName, String productType, Double price, boolean isVegan, int calories, int proteinAmount, String base) {
+    public FoodItem(String productName, String productType, BigDecimal price, boolean isVegan, int calories, int proteinAmount, String base) {
         super(productName, productType, price);
         this.isVegan = isVegan;
         this.calories = calories;
@@ -24,7 +26,7 @@ public class FoodItem extends Product {
         this.hasBase = true;
     }
 
-    public FoodItem(String productName, String productType, Double price) {
+    public FoodItem(String productName, String productType, BigDecimal price) {
         super(productName, productType, price);
         super.setProductModel("Food");
     }
